refactor: move paint_elevator_indicator to widgets.rs
Completes the renderer.rs split — all widget paint fns now live in
widgets.rs. renderer.rs drops to ~280 lines (pure orchestrator).
